Adding Visual Studio Project/Solution files
Just to make it easier to work on the project using Visual Studio.  Do
not think this means you have to have VS to work on this project.  Feel
free to add project files for any other IDEs as an easy ay to access all
of the JS/HTML/CSS/README files